requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
Faker==19.6.2
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.21
gunicorn==21.2.0
openai==1.3.8
langchain==0.0.350
langchain-openai==0.0.2
pydantic==2.5.0 